+++ src/cosine_annealing_with_restarts.py
@@ -2,7 +2,7 @@ import math
 
 from torch.optim.lr_scheduler import _LRScheduler
 
-from utils1 import to_string
+from utils import to_string
 
 
 class CosineAnnealingRestartingLR(_LRScheduler):
